took paths out of the code.
path to images and fonts directory is now provided by two symlinks created
in the directory where splash binary is supposed to be.

image or fontname (which are hardcoded) are joined with splash's path +
base name to form a name of needed object to cache.

if splash is copied to "/usr/bin/splash" and we are using file from images
folder with name picture01.jpg, let's create symlink with name "splash.fonts"
and "splash.images", put it into splash running dir and point to real
directories.

/usr/bin/splash + images/picture01.jpg = /usr/bin/splash.images/picture01.jpg
so
cd /usr/bin
ln -s [real_path_to_objects] ./splash.images

done!